langchain_sql_agent/requirements.txt
ibm_watsonx_ai
numpy==1.26.4
langchain==0.3.24
langchain_community==0.3.22
langchain_core==0.3.55
langchain_ibm==0.3.10
langchain-openai
langgraph==0.3.31
pandas==2.2.3
python-dotenv==1.1.0
SQLAlchemy==2.0.20
langchain-groq